vital_lab_extraction/vital_lab_extraction/requirements.txt
requests
boto3==1.34.34
botocore==1.34.34
jmespath==1.0.1
numpy==1.26.3
pandas==2.2.0
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
python-dateutil==2.8.2
pytz==2024.1
rapidfuzz==3.6.1
s3transfer==0.10.0
six==1.16.0
tzdata==2023.4
urllib3==2.0.7